Fix split package issue
